% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiowa_palette.R
\name{uiowa_cols}
\alias{uiowa_cols}
\title{Function to extract uiowa colors as hex codes}
\usage{
uiowa_cols(...)
}
\arguments{
\item{...}{Character names of uiowa_colors}
}
\description{
Function to extract uiowa colors as hex codes
}
