% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiowa_palette.R
\name{uiowa_pal}
\alias{uiowa_pal}
\title{Return function to interpolate a uiowa color palette}
\usage{
uiowa_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in uiowa_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a uiowa color palette
}
