% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drhur.R
\name{drhur}
\alias{drhur}
\title{Calling for an Exercise in Dr Hu's R Workshop}
\usage{
drhur(workshop, language = "cn")
}
\arguments{
\item{workshop}{A character vector indicating the name of the workshop to open. See details for the workshop list.}

\item{language}{A character vector indicating the display language. The current version includes workshops in English (\code{en}) and simple Chinese (\code{cn}). The default is Chinese.}
}
\description{
The function calls a \code{\link[learnr]{run_tutorial}} function to open a \code{learnr} project. The project list is continually updated along with the R workshop with Dr. Hu.
}
\details{
The list of workshop available for the current version:
\itemize{
\item \code{calculate}: Basic syntax of R (Chinese).
\item \code{play}: Basic data structure in R; inputting, outputting, and creating data.
\item \code{wrangle}: Generalizing (such as sorting and summarizing) and extracting (selecting or adjusting certain columns or rows) data frames (English and Chinese).
\item \code{describe}: Skills to describe variable and data.
}
}
